Create separate `Portal` model for Payment Portal info
Earlier, this info was stored as part of the user data. However,
decoupling it makes it easier for users to manage multiple payment
portals in the future, as well as perhaps add the ability for a
single portal to be managed by multiple users.

Additionally, this will help separate the user's identity with that
of the company; otherwise we would have had to post messages
everywhere telling the user to fill in their company's name, and
not their own, under "name", lest people get worried about paying
a Mr. XYZ instead of the So-and-So Company.