Make "invalid password" flash instead of coming as form error
The other way gives a 404 error in the live server when the password
is incorrect, for some reason. And plus, this way looks nicer. (Yes,
I know I should be changing the template to make form errors as nice
as flashes, but I'm too lazy to do that at the moment).